Add user's active games on player home
Fetch all active games of the logged in user in the view function and pass
it to the response as a list.
Active games are those where the user is a player and it is the player's
chance to move.

Display the active games in home.html using django template for loop.