Add player view method to accept an invitation
The method retrieves the Invitation object with the given id.
It checks if the current user is permitted to action the invitation.
For GET request, it renders the invitation template.
For POST request, if invitation accepted, starts the game and deletes
the invitation. If declined, deletes the invitation.
The user is redirected to the player home.