Add view method to make a move.
This method retrieves the game object based on the game id received.
It checks if it is the current user's move. If not - permission denied.
It then creates a new move from game model's method.
The form object is created based on the submitted data and the move object.
Form is validated and user is redirected based on the form validation.